view/eyes/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu

torch==2.2.2+cpu
torchvision==0.17.2+cpu
torchaudio
transformers==4.33.2
Pillow
open3d==0.13.0
omegaconf
hydra-core
iopath
contexttimer
decord
diffusers==0.16.0
easydict==1.9
einops>=0.4.1
fairscale==0.4.4
ftfy
h5py
moviepy
nltk
opencv-python-headless==4.5.5.64
opendatasets
peft
plotly
pre-commit
pycocoevalcap
pycocotools
python-magic
pyyaml-env-tag==0.1
scikit-image
sentencepiece
soundfile
spacy
streamlit
timm==0.4.12
webdataset
tokenizers
numpy==1.24.4

git+https://github.com/salesforce/LAVIS.git